global change in structure of pages and protected by permissions

+++ jirello/views.py
@@ -3,6 +3,7 @@ from jirello.models import User, Task, Sprint, ProjectModel
 from jirello.forms import RegistrationForm, AuthenticationForm, ProjectForm, SprintForm, TaskForm
 from django.http import HttpResponse, HttpResponseRedirect
 from django.shortcuts import get_object_or_404
+from django.http import Http404
 from django.core.urlresolvers import reverse
 
 from django.contrib.auth.decorators import login_required
@@ -11,7 +12,6 @@ from django.contrib.auth import logout as auth_logout
 from django.contrib.auth import login as auth_login
 from guardian.decorators import permission_required_or_403
 from guardian.shortcuts import assign_perm
-
 
 
 def register(request):
@@ -92,28 +92,17 @@ def new_project(request):
             form.save()
             # assign permissions for each user in cleaned data
             for u in form.cleaned_data['users']:
-                assign_perm('view_project', u, form.instance)
+                assign_perm('can_view', u, form.instance)
             return HttpResponseRedirect('/jirello/projects')
     context_dict = {'form': form, }
     return render(request, 'jirello/new_project.html', context_dict)
-
-
-def edit_project(request, projectmodel_id):
-    p = ProjectModel.objects.get(pk=projectmodel_id)
-    form = ProjectForm(instance=p)
-    if request.method == 'POST':
-        form = ProjectForm(request.POST or None, instance=p)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect('/jirello/projects')
-    return render(request, 'jirello/edit.html', {'form': form})
 
 
 def new_sprint(request, projectmodel_id):
     form = SprintForm
     if request.method == 'POST':
         form = SprintForm(request.POST)
-        if form.is_valid:
+        if form.is_valid():
             f = form.save(commit=False)
             f.owner = request.user
             f.project_id = projectmodel_id
@@ -122,18 +111,6 @@ def new_sprint(request, projectmodel_id):
                 'project_detail', args=[projectmodel_id, ]))
     context_dict = {'form': form, 'project_id': projectmodel_id, }
     return render(request, 'jirello/new_sprint.html', context_dict)
-
-
-def edit_sprint(request, projectmodel_id, sprint_id):
-    s = Sprint.objects.get(pk=sprint_id)
-    form = SprintForm(instance=s)
-    if request.method == 'POST':
-        form = SprintForm(request.POST or None, instance=s)
-        if form.is_valid():
-            form.save()
-            return HttpResponseRedirect(reverse(
-                'sprint_detail', args=[projectmodel_id, sprint_id]))
-    return render(request, 'jirello/edit.html', {'form': form})
 
 
 def new_task(request, projectmodel_id):
@@ -156,36 +133,73 @@ def new_task(request, projectmodel_id):
     return render(request, 'jirello/new_task.html', context_dict)
 
 
+def edit_project(request, projectmodel_id):
+    project = ProjectModel.objects.get(pk=projectmodel_id)
+    form = ProjectForm(instance=project)
+    if request.method == 'POST':
+        form = ProjectForm(request.POST or None, instance=project)
+        if request.POST.get('delete'):
+            project.delete()
+        return HttpResponseRedirect('/jirello/projects')
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect('/jirello/projects')
+    return render(request, 'jirello/edit.html', {'form': form})
+
+
+def edit_sprint(request, projectmodel_id, sprint_id):
+    sprint = Sprint.objects.get(pk=sprint_id)
+    form = SprintForm(instance=sprint)
+    if request.method == 'POST':
+        form = SprintForm(request.POST or None, instance=sprint)
+        if request.POST.get('delete'):
+            sprint.delete()
+            return HttpResponseRedirect(
+                reverse('project_detail', args=[projectmodel_id, ]))
+
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse(
+                'sprint_detail', args=[projectmodel_id, sprint_id]))
+    return render(request, 'jirello/edit.html', {'form': form})
+
+
 def edit_task(request):
     pass
 
 
-@permission_required_or_403('view_project',
+@permission_required_or_403('can_view',
                             (ProjectModel, 'pk', 'projectmodel_id'))
 def project_detail(request, projectmodel_id):
     # 404 error if project does not exist
-    get_object_or_404(ProjectModel, pk=projectmodel_id)
-    project = ProjectModel.objects.filter(
-        pk=projectmodel_id).prefetch_related('users', )
+    # get_object_or_404(ProjectModel, pk=projectmodel_id) !!!!!!!!!Problem
+    # with related objects
+    try:
+        project = ProjectModel.objects.filter(
+            pk=projectmodel_id).prefetch_related('users', )
+    except ProjectModel.DoesNotExist:
+        raise Http404("No project matches the given query.")
     # 'sprints__tasks'
     sprints = Sprint.objects.filter(
         project_id=projectmodel_id).order_by('date_end').prefetch_related('tasks')
     context_dict = {'project': project, 'sprints': sprints}
-
-    if request.POST.get('delete'):
-        project.delete()
-        return HttpResponseRedirect('/jirello/projects')
     return render(request, 'jirello/project_detail.html', context_dict)
 
 
+@permission_required_or_403('can_view',
+                            (ProjectModel, 'pk', 'projectmodel_id'))
 def sprint_detail(request, projectmodel_id, sprint_id):
     # 404 error if project does not exist
-    get_object_or_404(Sprint, pk=sprint_id)
-    sprint = Sprint.objects.get(pk=sprint_id)
-    context_dict = {'sprint': sprint, 'projectmodel_id': projectmodel_id}
-
-    if request.POST.get('delete'):
-        sprint.delete()
-        return HttpResponseRedirect(
-            reverse('project_detail', args=[projectmodel_id, ]))
+    sprint = get_object_or_404(Sprint, pk=sprint_id)
+    tasks = Task.objects.filter(sprints__id=sprint_id).order_by('storypoints')
+    context_dict = {'sprint': sprint, 'projectmodel_id': projectmodel_id, 'tasks':tasks}
     return render(request, 'jirello/sprint_detail.html', context_dict)
+
+
+@permission_required_or_403('can_view',
+                            (ProjectModel, 'pk', 'projectmodel_id'))
+def task_detail(request, projectmodel_id, task_id):
+    # 404 error if project does not exist
+    # get_object_or_404(Sprint, pk=sprint_id)
+    # task = Task.objects.get(pk=task_id)
+    pass
